requirements.txt
spotipy==2.19.0          # Spotify API library for Python
requests==2.28.2         # To make HTTP requests if needed
ffmpeg-python==0.2.0     # FFmpeg bindings for Python
pytube==12.1.0           # If downloading YouTube content is also needed
python-dotenv==1.0.0     # For managing environment variables (API keys)
googlesearch-python
colorama
getch
regex